+++ services/workflows-service/src/workflow-defintion/dtos/create-demo-workflow-eu-definition-dto.ts
@@ -1,0 +1,11 @@
+import { IsString } from 'class-validator';
+import { ApiProperty } from '@nestjs/swagger';
+
+export class CreateDemoWorkflowDefinitionEuDto {
+  @ApiProperty({
+    required: true,
+    type: String,
+  })
+  @IsString()
+  customerId!: string;
+}