+++ packages/config/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ballerine/config
 
+## 1.1.43
+
+### Patch Changes
+
+- bump
+
+## 1.1.42
+
+### Patch Changes
+
+- bump
+
+## 1.1.41
+
+### Patch Changes
+
+- bump
+
 ## 1.1.40
 
 ### Patch Changes